backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-generativeai
python-multipart
numpy
scikit-learn
torch
transformers
scipy
